- Refactor: removes routes from server.js and pulls them into their own route files - Refactor: updates UserController to add class with function headers for createAccount, login, getAllUsers
