+++ backend/src/controllers/UserController.js
@@ -1,0 +1,42 @@
+const User = require('../models/User');
+
+class UserController {
+    /** @type { User } */
+    user;
+
+    constructor() {
+        this.user = User;
+    }
+
+    /** PUBLIC METHODS */
+
+    /**
+     * Performs the account creation process based on the parameters passed to the func
+     * @param {string} username 
+     * @param {string} pwd 
+     * @returns A promise which resolves if the account creation succeeds, and rejects if it fails
+     */
+    createAccount(username, pwd) {
+
+    }
+
+    /**
+     * Performs authentication of a username and password
+     * @param {string} username 
+     * @param {string} pwd 
+     * @returns A promise which resolves to a login token if the login succeeds, and rejects if it fails
+     */
+    login(username, pwd) {
+
+    }
+
+    /**
+     * Gets the usernames for all the users in the Users table
+     * @returns An array of usernames
+     */
+    getAllUsers() {
+
+    }
+}
+
+module.exports = UserController;
